requirements.txt
python-dotenv==0.10.1
requests==2.20.1
beautifulsoup4==4.6.3
pdfminer.six==20181108
